Removes the test_util module
